// src/types/Category.ts
export enum Category {
  Tracking = "Tracking",
  DataCollection = "Data Collection",
  ThirdPartySharing = "Third-Party Sharing",
  Surveillance = "Surveillance",
  UserContentAndRights = "UserContentAndRights",
  DataRetention = "Data Retention",
  SecurityMeasures = "Security Measures",
  Cookies = "Cookies and Similar Technologies",
  DataTransfers = "Data Transfers and Storage Location",
  UserProfiling = "User Profiling and Automated Decision Making",
  DataBreachNotification = "Data Breach Notification",
}